altered commandline behaviour, made rasters transpose of before due to read and write performance issues

+++ src/cxx_tools/dynamic2darray.h
@@ -14,12 +14,12 @@
 #include "arrayvalidate.h"
 
 template<typename Type>
-Type** allocate2DArray(unsigned int numRows, unsigned int numCols)
+Type** allocate2DArray(unsigned long long numRows, unsigned long long numCols)
 {
 	Type** retArr = (Type **)calloc(numRows, sizeof(Type *));
 	retArr[0] = (Type *)calloc(numRows * numCols, sizeof(Type));
 
-	for (size_t i = 1; i < numRows; i++)
+	for (unsigned long long i = 1; i < numRows; i++)
 	{
 		retArr[i] = &(retArr[0][i * numCols]);
 	}
@@ -29,7 +29,7 @@ Type** allocate2DArray(unsigned int numRows, unsigned int numCols)
 
 template <typename Type>
 void write2DArray(std::string out_file_name, Type **inArr,
-	unsigned int num_row, unsigned int num_col, bool append = false)
+	unsigned long long num_row, unsigned long long num_col, bool append = false)
 {
 	std::ios_base::openmode app_opt = (append) ? std::ios_base::app : (std::ios_base::openmode)0;
 	std::fstream out_file_buf(out_file_name.c_str(), std::ios::out | std::ios::binary | app_opt);
@@ -51,7 +51,7 @@ void delete2DArray(Type** array)
 }
 
 template <typename Type>
-inline bool validate2DArray(Type **array, unsigned int numElements)
+inline bool validate2DArray(Type **array, unsigned long long numElements)
 {
 	return validateFloatArray(array[0], numElements);
 }
